Renames camera package
