+++ src/pyrite/_camera/camera.py
