+++ src/pyrite/_camera/chase_camera.py
