+++ src/pyrite/game.py
@@ -13,7 +13,7 @@ from .core.render_system import RenderSystem, RenderManager
 from .core.rate_settings import RateSettings
 from .core.system_manager import SystemManager
 
-from .camera.camera import Camera
+from pyrite._camera.camera import Camera
 from .rendering import OrthoProjection, Viewport
 from .services import CameraService
 from .transform import transform_system
